requirements.txt
nglview
numpy
pandas
Pillow
pytest
Requests
scikit_learn==1.1.2
scipy
torch
tqdm
transformers==4.24.0